refactor(scheduling.py): move some functions from utils.py to scheduling.py
